+++ backend/eslint.config.js
@@ -1,0 +1,47 @@
+import pluginJs from "@eslint/js";
+import globals from "globals";
+import tseslint from "typescript-eslint";
+import eslintConfigPrettier from "eslint-config-prettier";
+
+export default [
+  {
+    languageOptions: { globals: globals.node },
+  },
+  {
+    ignores: [
+      "node_modules/*",
+      "node_modules",
+      "node_modules/**/*",
+      "dist/*",
+      "dist/**/*",
+      "dist",
+      ".database",
+      ".database/*",
+    ],
+  },
+  pluginJs.configs.recommended,
+  ...tseslint.configs.strictTypeChecked,
+  eslintConfigPrettier,
+  {
+    rules: {
+      "@typescript-eslint/no-unused-vars": [
+        "error",
+        {
+          args: "all",
+          argsIgnorePattern: "^_",
+          caughtErrors: "all",
+          caughtErrorsIgnorePattern: "^_",
+          destructuredArrayIgnorePattern: "^_",
+          varsIgnorePattern: "^_",
+          ignoreRestSiblings: true,
+        },
+      ],
+    },
+    languageOptions: {
+      parserOptions: {
+        projectService: true,
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
+  },
+];